Refactored shared contracts into lib directory
I moved the most obviously shared contracts into a `lib` directory. I
moved the tests accordingly.

In the `test` directory, `lib` now contains tests for the
`contracts/lib` files. The test libraries have been moved to
`test/test_lib`.

Fixes 303
