+++ contracts/gateway/BrandedToken.sol
@@ -22,7 +22,7 @@ pragma solidity ^0.4.23;
 //
 // ----------------------------------------------------------------------------
 
-import "./SafeMath.sol";
+import "../lib/SafeMath.sol";
 
 /** utility chain contracts */
 import "./EIP20Token.sol";
