+++ contracts/gateway/EIP20Token.sol
@@ -22,7 +22,7 @@ pragma solidity ^0.4.23;
 // ----------------------------------------------------------------------------
 
 import "./EIP20Interface.sol";
-import "./SafeMath.sol";
+import "../lib/SafeMath.sol";
 
 /**
  *  @title EIP20Token contract which implements EIP20Interface.
