+++ contracts/lib/Hasher.sol
