+++ test/gateway/Core.js
@@ -20,11 +20,11 @@
 // ----------------------------------------------------------------------------
 
 const coreUtils = require('./Core_utils.js')
-    , utils = require('../lib/utils.js')
+    , utils = require('../test_lib/utils.js')
     , proof = require('../data/proof')
     , RLP = require('rlp')
     , BigNumber = require('bignumber.js')
-    , web3EventsDecoder = require('../lib/event_decoder.js')
+    , web3EventsDecoder = require('../test_lib/event_decoder.js')
 ;
 
 contract('Core', function (accounts) {
