+++ test/gateway/EIP20Token_protocol_extensions.js
@@ -20,7 +20,7 @@
 // ----------------------------------------------------------------------------
 
 const BigNumber = require('bignumber.js');
-const Utils = require('../lib/utils.js');
+const Utils = require('../test_lib/utils.js');
 const EIP20Token_utils = require('./EIP20Token_utils.js');
 
 ///
