+++ test/gateway/OpenSTUtility_utils.js
@@ -21,7 +21,7 @@
 
 const BigNumber = require('bignumber.js');
 const rootPrefix = "../.."
-  , constants = require(rootPrefix + '/test/lib/constants')
+  , constants = require(rootPrefix + '/test/test_lib/constants')
 ;
 
 var OpenSTUtility = artifacts.require("./OpenSTUtilityMock.sol");
