+++ test/gateway/OpenSTValue_utils.js
@@ -22,7 +22,7 @@
 const BigNumber = require('bignumber.js');
 
 const rootPrefix = "../.."
-  , constants = require(rootPrefix + '/test/lib/constants')
+  , constants = require(rootPrefix + '/test/test_lib/constants')
 ;
 
 //var OpenSTValue = artifacts.require("./OpenSTValue.sol");
