+++ test/gateway/Registrar.js
@@ -20,8 +20,8 @@
 // ----------------------------------------------------------------------------
 
 const BigNumber       = require('bignumber.js');
-const Utils           = require('../lib/utils.js');
-const HashLock        = require('../lib/hash_lock.js');
+const Utils           = require('../test_lib/utils.js');
+const HashLock        = require('../test_lib/hash_lock.js');
 const Registrar_utils = require('./Registrar_utils.js');
 
 ///
