+++ test/gateway/Registrar_utils.js
@@ -22,7 +22,7 @@
 const BigNumber = require('bignumber.js');
 
 const rootPrefix = "../.."
-  , constants = require(rootPrefix + '/test/lib/constants')
+  , constants = require(rootPrefix + '/test/test_lib/constants')
 ;
 
 var MockToken = artifacts.require("./MockToken.sol");
