+++ test/gateway/SimpleStake_integrated.js
@@ -21,7 +21,7 @@
 
 const Assert = require('assert');
 const BigNumber = require('bignumber.js');
-const Utils = require('../lib/utils.js');
+const Utils = require('../test_lib/utils.js');
 const SimpleStakeUtils = require('./SimpleStake_utils.js')
 
 ///
