+++ test/test_lib/hash_lock.js
